# Conflicts:
#	MacOSX/English.lproj/InfoPlist.strings
#	MacOSX/Info.plist
#	MacOSX/QuakeSpasm.xcodeproj/project.pbxproj
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_revision.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_stdinc.h
#	MacOSX/SDL2.framework/Versions/A/SDL2
#	MacOSX/codecs/include/xmp.h
#	MacOSX/codecs/lib/libmpg123.dylib
#	MacOSX/codecs/lib/libopusfile.dylib
#	MacOSX/codecs/lib/libxmp.dylib
#	Misc/pak/default.cfg.diff
#	Misc/pak/default.cfg.orig
#	Misc/pak/maps/e1m1.diff
#	Misc/pak/maps/e1m1.ent.orig
#	Misc/pak/maps/e1m2.diff
#	Misc/pak/maps/e1m2.ent.orig
#	Misc/pak/maps/e1m4.diff
#	Misc/pak/maps/e1m4.ent.orig
#	Misc/pak/maps/e2m2.diff
#	Misc/pak/maps/e2m2.ent.orig
#	Misc/pak/maps/e2m3.diff
#	Misc/pak/maps/e2m3.ent.orig
#	Misc/pak/maps/e2m7.diff
#	Misc/pak/maps/e2m7.ent.orig
#	Quake/Makefile
#	Quake/Makefile.darwin
#	Quake/Makefile.w32
#	Quake/Makefile.w64
#	Quake/Makefile.wat
#	Quake/cl_demo.c
#	Quake/cl_main.c
#	Quake/cmd.c
#	Quake/common.c
#	Quake/common.h
#	Quake/gl_mesh.c
#	Quake/gl_model.c
#	Quake/gl_refrag.c
#	Quake/gl_rlight.c
#	Quake/gl_rmain.c
#	Quake/gl_rmisc.c
#	Quake/gl_texmgr.c
#	Quake/gl_texmgr.h
#	Quake/gl_vidsdl.c
#	Quake/gl_warp.c
#	Quake/glquake.h
#	Quake/host.c
#	Quake/host_cmd.c
#	Quake/image.c
#	Quake/in_sdl.c
#	Quake/menu.c
#	Quake/pr_cmds.c
#	Quake/progs.h
#	Quake/r_alias.c
#	Quake/r_brush.c
#	Quake/r_part.c
#	Quake/r_world.c
#	Quake/render.h
#	Quake/sbar.c
#	Quake/server.h
#	Quake/snd_wave.c
#	Quake/sv_main.c
#	Quake/sv_phys.c
#	Quake/sys_sdl_unix.c
#	Windows/VS2005/quakespasm-sdl2.vcproj
#	Windows/VS2005/quakespasm.vcproj
#	Windows/VisualStudio/ironwail.sln

+++ Quake/progs.h
@@ -330,7 +330,7 @@ int SAVE_NUM_FOR_EDICT (savedata_t *save, edict_t *e);
 
 #define	NEXT_EDICT(e)		((edict_t *)( (byte *)e + qcvm->edict_size))
 
-#define	EDICT_TO_PROG(e)	((byte *)e - (byte *)qcvm->edicts)
+#define	EDICT_TO_PROG(e)	(int)((byte *)e - (byte *)qcvm->edicts)
 #define PROG_TO_EDICT(e)	((edict_t *)((byte *)qcvm->edicts + e))
 #define SAVE_PROG_TO_EDICT(s, e)	((edict_t *)((byte *)s->edicts + e))
